requirements.txt
flask~=3.1.0
python-dotenv~=1.1.0
pyyaml~=6.0.2
requests
python-dotenv
aiogram~=3.20.0.post0